Safari-611.1.10/LayoutTests/canvas/philip/tests/size.attributes.style.html

<!DOCTYPE html>
<title>Canvas test: size.attributes.style</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<body>
<p id="passtext">Pass</p>
<p id="failtext">Fail</p>
<p class="output">These images should be identical:</p>
<canvas id="c" class="output" width="50" height="30" style="width: 100px; height: 50px"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:<p><img src="size.attributes.style.png" class="output expected" id="expected" alt="">
<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

_assertEqual(canvas.width, 50, "canvas.width", "50");
_assertEqual(canvas.height, 30, "canvas.height", "30");


});
</script>
