LayoutTests/canvas/philip/tests/size.attributes.style-expected.txt

< [index] >
size.attributes.style
Canvas size is independent of CSS resizing
References: size.css
Actual output:
Expected output:

Passed
